// packages/frontend/src/generated/types/UserResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `user` entity.
 */
export interface UserResponse {
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string,null
   */
  fullname?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string
   */
  role: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
  /**
   * @type string
   */
  username: string
}